+++ lua/rustaceanvim/neotest/trans.lua
@@ -1,0 +1,57 @@
+local M = {}
+
+---@param runnable RARunnable
+---@return string | nil
+local function get_test_path(runnable)
+  local executableArgs = runnable.args and runnable.args.executableArgs or {}
+  return #executableArgs > 0 and executableArgs[1] or nil
+end
+
+---@param file_path string
+---@param runnable RARunnable
+---@return string
+function M.get_position_id(file_path, runnable)
+  local test_path = get_test_path(runnable)
+  return test_path and table.concat(vim.list_extend({ file_path }, { test_path }), '::') or file_path
+end
+
+---@param file_path string
+---@param runnable RARunnable
+---@return rustaceanvim.neotest.Position | nil
+function M.runnable_to_position(file_path, runnable)
+  local cargoArgs = runnable.args and runnable.args.cargoArgs or {}
+  if #cargoArgs > 0 and vim.startswith(cargoArgs[1], 'test') then
+    ---@type neotest.PositionType
+    local type
+    if vim.startswith(runnable.label, 'cargo test -p') then
+      type = 'dir'
+    elseif vim.startswith(runnable.label, 'test-mod') then
+      type = 'namespace'
+    elseif vim.startswith(runnable.label, 'test') then
+      type = 'test'
+    else
+      return
+    end
+    local location = runnable.location
+    local start_row, start_col, end_row, end_col = 0, 0, 0, 0
+    if location then
+      start_row = location.targetRange.start.line + 1
+      start_col = location.targetRange.start.character
+      end_row = location.targetRange['end'].line + 1
+      end_col = location.targetRange['end'].character
+    end
+    local test_path = get_test_path(runnable)
+    ---@type rustaceanvim.neotest.Position
+    local pos = {
+      id = M.get_position_id(file_path, runnable),
+      name = test_path or runnable.label,
+      type = type,
+      path = file_path,
+      range = { start_row, start_col, end_row, end_col },
+      runnable = runnable,
+    }
+    return pos
+  end
+end
+
+return M