+++ CHANGELOG.md
@@ -6,6 +6,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.4.0] - 2024-01-31
+
+### Added
+
+- You can now register a `rustaceanvim.neotest` adapter with [neotest](https://github.com/nvim-neotest/neotest).
+  It will query rust-analyzer for test positions and test commands in any
+  buffer to which the LSP client is attached.
+  If you do so, `tools.test_executor` will default to a new `'neotest'`
+  executor, which will use neotest to run `testables` or `runnables` that are tests.
+
 ## [4.3.0] - 2024-01-31
 
 ### Changed